renamed *-desktop & *-mobile to have deviec agnostic names
