--- libs/designsystem/list/src/list-item-mobile/list-item-mobile.component.scss
@@ -1,24 +1,0 @@
-@use '@kirbydesign/core/src/scss/utils';
-
-$divider-color: utils.get-color('background-color');
-
-:host {
-  overflow: hidden;
-}
-
-:host-context(.has-divider) {
-  ion-item-sliding:not(.last) {
-    border-bottom: 1px solid $divider-color;
-  }
-}
-
-// .item_container {
-//   display: flex;
-//   flex-direction: row;
-//   align-items: center;
-//   justify-content: space-between;
-//   flex-wrap: nowrap;
-//   // width: 100%;
-//   // height: 100%;
-//   // padding: size('xxxs') size('s');
-// }
