+++ libs/designsystem/list/src/list-item/list-item.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, Output, TemplateRef } from '@angular/core';
 import { PlatformService, ThemeColor } from '@kirbydesign/designsystem/helpers';
-import { BoundaryClass } from '../list-item-mobile/list-item-mobile.component';
+import { BoundaryClass } from '../list-item-swipe/list-item-swipe.component';
 import { ListSwipeAction } from '../list-swipe-action.type';
 
 export type DeviceType = 'desktop' | 'mobile';
