+++ libs/designsystem/list/src/list.component.ts
@@ -16,7 +16,7 @@ import { ThemeColor } from '@kirbydesign/core';
 
 import { InfiniteScrollDirective } from './directives/infinite-scroll.directive';
 import { ListHelper } from './helpers/list-helper';
-import { BoundaryClass } from './list-item-mobile/list-item-mobile.component';
+import { BoundaryClass } from './list-item-swipe/list-item-swipe.component';
 
 import { ListSwipeAction } from './list-swipe-action.type';
 import {
